requirements.txt
transformers==4.21.2
tokenizers==0.11.6
cached-property
networkx
penman>=1.1.0
pytorch-ignite
regex
sacrebleu
smatch
wandb
PyYAML>=5.1
matplotlib
pandas
torch-geometric
spacy